protocol/inspircd12: Remove support
