+++ data/example.conf
@@ -250,7 +250,6 @@ serverinfo
  *  - bahamut
  *  - charybdis
  *  - hybrid
- *  - inspircd12
  *  - inspircd20
  *  - ngircd
  *  - plexus
