+++ docs/README
@@ -171,7 +171,7 @@ Table of Contents
         * Bahamut 1.4.27 or later (including 1.8)
         * Charybdis 3.4 or later
         * Hybrid 8.1 or later
-        * InspIRCd 1.2 or 2.0
+        * InspIRCd 2.0
         * ngIRCd 19.2 or later
         * Plexus 3 or later
         * Ratbox 2.0.6 or later
